Get rid of no longer functioning test and rename index.ts --> app.ts
